some minor code fixes and refactorings
